+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [3.0.4] - 2022-06-23
+### Added
+- Add stats page for every team member
+
+
+## [3.0.3] - 2022-06-21
+### Added
+- Add zendesk performance graphs for the main dashboard
+
+
+## [3.0.2] - 2022-06-20
+### Added
+- Graphs for the main dashboard
+
+### Changed
+- Stats cards 
+
+
 ## [3.0.1] - 2022-06-18
 ### Changed
 - Add balance to zendesk task distribution algorithms